+++ blitz/src/buffers/index_buffer.rs
@@ -22,7 +22,7 @@ pub struct IndexBuffer {
 
 impl IndexBuffer {
     pub unsafe fn new(instance: &Instance, device: &Device, indices: &[u16]) -> Result<Self> {// Size
-        let size = (size_of::<u16>() *indices.len()) as u32;
+        let size = (size_of::<u16>() *indices.len()) as u64;
 
         // Buffer
